refactor socket handler
