+++ backend/logic/avatarHandling.js
@@ -1,0 +1,14 @@
+import { lobbies } from './lobbyManagement.js';
+
+export function registerAvatarHandlers(io, socket, avatarFiles) {
+  socket.on('change-avatar', (gameCode) => {
+    const lobby = lobbies[gameCode];
+    if (!lobby) return;
+    const player = socket.data.playerName;
+    const avatars = lobby.avatars;
+    if (!player || !avatars) return;
+    const file = avatarFiles[Math.floor(Math.random() * avatarFiles.length)];
+    avatars[player] = file;
+    io.to(gameCode).emit('avatar-changed', { player, file });
+  });
+}
